+++ skquantum/combinatorial/__init__.py
@@ -1,0 +1,1 @@
+from .weighted_maxcut import WeightedMaxCut
